Fix stogit url bug
