+++ stograde/toolkit/__main__.py
@@ -9,12 +9,14 @@ import os.path
 import logging
 
 from .ci_analyze import ci_analyze
+from .download_specs import create_data_dir
+from .stogit_url import compute_stogit_url
 from ..student import clone_student
 from ..common import chdir, run
 from ..specs import load_all_specs, check_dependencies, check_architecture
 from .find_update import update_available
 from .process_student import process_student
-from .args import process_args, compute_stogit_url
+from .args import process_args
 from .progress_bar import progress_bar
 from .save_recordings import save_recordings, gist_recordings
 from .tabulate import tabulate
@@ -49,27 +51,6 @@ def run_server(basedir, port):
     return
 
 
-def download_specs(course, basedir, stogit):
-    spec_urls = {
-        'sd': 'https://github.com/StoDevX/cs251-specs.git',
-        'hd': 'https://github.com/StoDevX/cs241-specs.git',
-        'ads': 'https://github.com/StoDevX/cs253-specs.git',
-        'os': 'https://github.com/StoDevX/cs273-specs.git'
-    }
-    course = course.split("/")[0].lower()
-    try:
-        url = spec_urls[course]
-    except KeyError:
-        print("Course {} not recognized".format(course))
-        sys.exit(1)
-    with chdir(basedir):
-        print('Downloading specs for {}'.format(course.upper()))
-        run(['git', 'clone', url, 'data'])
-        print('Download complete')
-        if not stogit:
-            return compute_stogit_url(course=course, stogit=None, _now=datetime.date.today())
-
-
 def run_analysis(*, no_progress=False, parallel, single_analysis, usernames, workers=1):
     results = []
     records = []
@@ -95,7 +76,7 @@ def run_analysis(*, no_progress=False, parallel, single_analysis, usernames, wor
 
 def main():
     basedir = getcwd()
-    args, usernames, assignments, stogit_url = process_args()
+    args, usernames, assignments = process_args()
     ci = args['ci']
     clean = args['clean']
     course = args['course']
@@ -130,33 +111,9 @@ def main():
         logging.debug('Checking out {}'.format(date))
 
     if not os.path.exists("data"):
-        if ci:
-            if course:
-                url = download_specs(course, basedir, stogit)
-                if not stogit:
-                    stogit_url = url
-            else:
-                print("data directory not found and no course specified")
-                sys.exit(1)
+        create_data_dir(ci, course, basedir)
 
-        else:
-            print('data directory not found', file=sys.stderr)
-            if course:
-                url = download_specs(course, basedir, stogit)
-                if not stogit:
-                    stogit_url = url
-            else:
-                download = input("Download specs? (Y/N)")
-                if download and download.lower()[0] == "y":
-                    repo = input("Which class? (SD/HD/ADS/OS)")
-                    if repo:
-                        url = download_specs(repo, basedir, stogit)
-                        if not stogit:
-                            stogit_url = url
-                    else:
-                        sys.exit(1)
-                else:
-                    sys.exit(1)
+    stogit_url = compute_stogit_url(stogit=stogit, course=course, _now=datetime.date.today())
 
     if re_cache_specs:
         try:
@@ -166,7 +123,7 @@ def main():
 
     specs = load_all_specs(basedir=os.path.join(basedir, 'data'), skip_update_check=skip_update_check)
     if not specs:
-        print('no specs loaded!')
+        print('No specs loaded!')
         sys.exit(1)
 
     if assignments:
